+++ plastron-cli/src/plastron/cli/commands/importcommand.py
@@ -6,7 +6,7 @@ from typing import TextIO
 
 from plastron.cli.commands import BaseCommand
 from plastron.client import Client
-from plastron.jobs.imports import ImportJob
+from plastron.jobs.importjob import ImportJob
 from plastron.models import get_model_class, ModelClassNotFoundError
 from plastron.rdf import uri_or_curie
 from plastron.utils import datetimestamp
