+++ plastron-repo/tests/jobs/test_import_job.py
@@ -1,4 +1,4 @@
-from plastron.jobs.imports import ImportJob
+from plastron.jobs.importjob import ImportJob
 
 
 def test_safe_job_id():
